requirements.txt
numpy==1.23.0
pandas==1.4.2
matplotlib==3.6.0
scikit-learn==1.0.2
Pillow==9.0.1
torch==1.13.0  
torchvision==0.14.0  
prettytable==2.4.0  
tqdm==4.64.0  

fastprogress==0.2.10
